refactor: extract util functions for safeSDK
